tests/wpt/tests/css/filter-effects/parsing/lighting-color-computed.html

<!DOCTYPE html>
<html>
<head>
<meta charset="utf-8">
<title>Filter Effects Module Level 1: getComputedStyle().lightingColor</title>
<link rel="help" href="https://drafts.fxtf.org/filter-effects/#LightingColorProperty">
<meta name="assert" content="lighting-color computed value is as specified.">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/css/support/computed-testcommon.js"></script>
</head>
<body>
<div id="target"></div>
<script>
test_computed_value("lighting-color", "rgb(1, 2, 3)");
</script>
</body>
</html>
